Rename import paths
