+++ desktop/apps/auction/__tests__/utils/renderTestComponent.js
@@ -3,8 +3,8 @@ import Articles from 'desktop/collections/articles.coffee'
 import Auction from 'desktop/models/auction.coffee'
 import CurrentUser from 'desktop/models/current_user.coffee'
 import React from 'react'
-import auctions from 'desktop/apps/auction2/reducers'
-import bootstrap from 'desktop/apps/auction2/__tests__/fixtures/auction'
+import auctions from 'desktop/apps/auction/reducers'
+import bootstrap from 'desktop/apps/auction/__tests__/fixtures/auction'
 import thunk from 'redux-thunk'
 import { Provider } from 'react-redux'
 import { applyMiddleware, createStore } from 'redux'
