+++ desktop/apps/auction/client.js
@@ -1,11 +1,11 @@
-import App from 'desktop/apps/auction2/components/App'
+import App from 'desktop/apps/auction/components/App'
 import Articles from 'desktop/collections/articles.coffee'
 import Auction from 'desktop/models/auction.coffee'
 import CurrentUser from 'desktop/models/current_user.coffee'
 import React from 'react'
 import ReactDOM from 'react-dom'
-import auctionReducer from 'desktop/apps/auction2/reducers'
-import configureStore from 'desktop/apps/auction2/utils/configureStore'
+import auctionReducer from 'desktop/apps/auction/reducers'
+import configureStore from 'desktop/apps/auction/utils/configureStore'
 import { rehydrateClient } from 'desktop/components/react/utils/renderReactLayout'
 
 // Rehydrate data from Server
