+++ desktop/apps/auction/components/App.js
@@ -1,5 +1,5 @@
-import DOM from 'desktop/apps/auction2/components/DOM'
-import Layout from 'desktop/apps/auction2/components/Layout'
+import DOM from 'desktop/apps/auction/components/DOM'
+import Layout from 'desktop/apps/auction/components/Layout'
 import PropTypes from 'prop-types'
 import React, { Component } from 'react'
 import ResponsiveWindow from 'desktop/components/react/responsive_window'
