+++ desktop/apps/auction/components/artwork_browser/__tests__/ArtworkBrowser.test.js
@@ -1,14 +1,14 @@
-import * as actions from 'desktop/apps/auction2/actions/artworkBrowser'
-import ArtistFilter from 'desktop/apps/auction2/components/artwork_browser/sidebar/ArtistFilter'
-import GridArtwork from 'desktop/apps/auction2/components/artwork_browser/main/artwork/GridArtwork'
-import ListArtwork from 'desktop/apps/auction2/components/artwork_browser/main/artwork/ListArtwork'
-import FilterSort from 'desktop/apps/auction2/components/artwork_browser/header/FilterSort'
-import MediumFilter from 'desktop/apps/auction2/components/artwork_browser/sidebar/MediumFilter'
-import RangeSlider from 'desktop/apps/auction2/components/artwork_browser/sidebar/RangeSlider'
-import Sidebar from 'desktop/apps/auction2/components/artwork_browser/sidebar/Sidebar'
-import auctions from 'desktop/apps/auction2/reducers'
+import * as actions from 'desktop/apps/auction/actions/artworkBrowser'
+import ArtistFilter from 'desktop/apps/auction/components/artwork_browser/sidebar/ArtistFilter'
+import GridArtwork from 'desktop/apps/auction/components/artwork_browser/main/artwork/GridArtwork'
+import ListArtwork from 'desktop/apps/auction/components/artwork_browser/main/artwork/ListArtwork'
+import FilterSort from 'desktop/apps/auction/components/artwork_browser/header/FilterSort'
+import MediumFilter from 'desktop/apps/auction/components/artwork_browser/sidebar/MediumFilter'
+import RangeSlider from 'desktop/apps/auction/components/artwork_browser/sidebar/RangeSlider'
+import Sidebar from 'desktop/apps/auction/components/artwork_browser/sidebar/Sidebar'
+import auctions from 'desktop/apps/auction/reducers'
 import React from 'react'
-import renderTestComponent from 'desktop/apps/auction2/__tests__/utils/renderTestComponent'
+import renderTestComponent from 'desktop/apps/auction/__tests__/utils/renderTestComponent'
 import { createStore } from 'redux'
 
 describe('auction/components/artwork_browser/ArtworkBrowser.test.js', () => {
