+++ desktop/apps/auction/components/artwork_browser/header/__tests__/HeaderMobile.test.js
@@ -1,7 +1,7 @@
-import renderTestComponent from 'desktop/apps/auction2/__tests__/utils/renderTestComponent'
-import HeaderMobile from 'desktop/apps/auction2/components/artwork_browser/header/HeaderMobile'
+import renderTestComponent from 'desktop/apps/auction/__tests__/utils/renderTestComponent'
+import HeaderMobile from 'desktop/apps/auction/components/artwork_browser/header/HeaderMobile'
 import sinon from 'sinon'
-import { __RewireAPI__ as RoutesRewireApi } from 'desktop/apps/auction2/actions/artworkBrowser'
+import { __RewireAPI__ as RoutesRewireApi } from 'desktop/apps/auction/actions/artworkBrowser'
 
 describe('auction/components/artwork_browser/header/HeaderMobile.test', () => {
   describe('<HeaderMobile />', () => {
