+++ desktop/apps/auction/components/artwork_browser/main/ArtworkDisplay.js
@@ -1,15 +1,15 @@
 import InfiniteScroll from 'desktop/components/react/infinite_scroll/InfiniteScroll'
 import Jump from 'desktop/components/jump/react'
-import MasonryArtwork from 'desktop/apps/auction2/components/artwork_browser/main/artwork/MasonryArtwork'
-import GridArtwork from 'desktop/apps/auction2/components/artwork_browser/main/artwork/GridArtwork'
-import ListArtwork from 'desktop/apps/auction2/components/artwork_browser/main/artwork/ListArtwork'
-import LoadingSpinner from 'desktop/apps/auction2/components/artwork_browser/main/LoadingSpinner'
+import MasonryArtwork from 'desktop/apps/auction/components/artwork_browser/main/artwork/MasonryArtwork'
+import GridArtwork from 'desktop/apps/auction/components/artwork_browser/main/artwork/GridArtwork'
+import ListArtwork from 'desktop/apps/auction/components/artwork_browser/main/artwork/ListArtwork'
+import LoadingSpinner from 'desktop/apps/auction/components/artwork_browser/main/LoadingSpinner'
 import MasonryGrid from 'desktop/components/react/masonry_grid/MasonryGrid'
 import PropTypes from 'prop-types'
 import React, { Component } from 'react'
 import block from 'bem-cn'
 import get from 'lodash.get'
-import { infiniteScroll } from 'desktop/apps/auction2/actions/artworkBrowser'
+import { infiniteScroll } from 'desktop/apps/auction/actions/artworkBrowser'
 import { connect } from 'react-redux'
 
 class ArtworkDisplay extends Component {
