+++ desktop/apps/auction/components/artwork_browser/sidebar/Sidebar.js
@@ -1,6 +1,6 @@
-import ArtistFilter from 'desktop/apps/auction2/components/artwork_browser/sidebar/ArtistFilter'
-import MediumFilter from 'desktop/apps/auction2/components/artwork_browser/sidebar/MediumFilter'
-import RangeSlider from 'desktop/apps/auction2/components/artwork_browser/sidebar/RangeSlider'
+import ArtistFilter from 'desktop/apps/auction/components/artwork_browser/sidebar/ArtistFilter'
+import MediumFilter from 'desktop/apps/auction/components/artwork_browser/sidebar/MediumFilter'
+import RangeSlider from 'desktop/apps/auction/components/artwork_browser/sidebar/RangeSlider'
 import PropTypes from 'prop-types'
 import React from 'react'
 import block from 'bem-cn'
