+++ desktop/apps/auction/components/artwork_browser/sidebar/__tests__/MediumFilter.test.js
@@ -1,6 +1,6 @@
 import React from 'react'
-import renderTestComponent from 'desktop/apps/auction2/__tests__/utils/renderTestComponent'
-import { test, __RewireAPI__ } from 'desktop/apps/auction2/components/artwork_browser/sidebar/MediumFilter'
+import renderTestComponent from 'desktop/apps/auction/__tests__/utils/renderTestComponent'
+import { test, __RewireAPI__ } from 'desktop/apps/auction/components/artwork_browser/sidebar/MediumFilter'
 
 const { MediumFilter } = test
