+++ desktop/apps/auction/components/layout/auction_info/__tests__/Registration.test.js
@@ -1,5 +1,5 @@
-import renderTestComponent from 'desktop/apps/auction2/__tests__/utils/renderTestComponent'
-import { test } from 'desktop/apps/auction2/components/layout/auction_info/Registration'
+import renderTestComponent from 'desktop/apps/auction/__tests__/utils/renderTestComponent'
+import { test } from 'desktop/apps/auction/components/layout/auction_info/Registration'
 
 const { Registration } = test
