+++ desktop/apps/auction/reducers/index.js
@@ -1,5 +1,5 @@
-import app, { initialState as appInitialState } from 'desktop/apps/auction2/reducers/app'
-import filter, { initialState as filterInitialState } from 'desktop/apps/auction2/reducers/artworkBrowser'
+import app, { initialState as appInitialState } from 'desktop/apps/auction/reducers/app'
+import filter, { initialState as filterInitialState } from 'desktop/apps/auction/reducers/artworkBrowser'
 import { combineReducers } from 'redux'
 import { composeReducers, responsiveWindowReducer } from 'desktop/components/react/responsive_window'
