+++ desktop/apps/auction/routes.js
@@ -1,19 +1,19 @@
-import * as actions from 'desktop/apps/auction2/actions/artworkBrowser'
-import App from 'desktop/apps/auction2/components/App'
+import * as actions from 'desktop/apps/auction/actions/artworkBrowser'
+import App from 'desktop/apps/auction/components/App'
 import Articles from 'desktop/collections/articles.coffee'
-import ArticlesQuery from 'desktop/apps/auction2/utils/queries/articles'
+import ArticlesQuery from 'desktop/apps/auction/utils/queries/articles'
 import Auction from 'desktop/models/auction.coffee'
-import MeQuery from 'desktop/apps/auction2/utils/queries/me'
+import MeQuery from 'desktop/apps/auction/utils/queries/me'
 import React from 'react'
-import SaleQuery from 'desktop/apps/auction2/utils/queries/sale'
-import auctionReducer from 'desktop/apps/auction2/reducers'
-import configureStore from 'desktop/apps/auction2/utils/configureStore'
-import footerItems from 'desktop/apps/auction2/utils/footerItems'
+import SaleQuery from 'desktop/apps/auction/utils/queries/sale'
+import auctionReducer from 'desktop/apps/auction/reducers'
+import configureStore from 'desktop/apps/auction/utils/configureStore'
+import footerItems from 'desktop/apps/auction/utils/footerItems'
 import get from 'lodash.get'
 import metaphysics from 'lib/metaphysics.coffee'
 import u from 'updeep'
-import { initialState as appInitialState } from 'desktop/apps/auction2/reducers/app'
-import { initialState as auctionWorksInitialState } from 'desktop/apps/auction2/reducers/artworkBrowser'
+import { initialState as appInitialState } from 'desktop/apps/auction/reducers/app'
+import { initialState as auctionWorksInitialState } from 'desktop/apps/auction/reducers/artworkBrowser'
 import { getLiveAuctionUrl } from 'utils/domain/auctions/urls'
 import { renderReactLayout } from 'desktop/components/react/utils/renderReactLayout'
