+++ desktop/apps/auction/server.js
@@ -7,5 +7,5 @@ const app = module.exports = express.Router()
 app.get('/sale2/:id', adminOnly, routes.index)
 app.get('/sale2/:id/confirm-registration', adminOnly, routes.index)
 
-app.get('/auction2/:id', adminOnly, routes.index)
-app.get('/auction2/:id/confirm-registration', routes.redirectLive, adminOnly, routes.index)
+app.get('/auction/:id', adminOnly, routes.index)
+app.get('/auction/:id/confirm-registration', routes.redirectLive, adminOnly, routes.index)
