+++ desktop/apps/auctions2/server.js
@@ -6,4 +6,4 @@ const app = module.exports = express.Router()
 
 app.get('/auctions2', adminOnly, routes.index)
 app.get('/auctions2/reminders', adminOnly, routes.reminders)
-app.get('/auction2', adminOnly, routes.redirectAuction)
+app.get('/auction', adminOnly, routes.redirectAuction)
