+++ desktop/index.js
@@ -9,7 +9,7 @@ app.use(require('./apps/auctions2'))
 app.use(require('./apps/artist'))
 app.use(require('./apps/artists'))
 app.use(require('./apps/auction'))
-app.use(require('./apps/auction2'))
+app.use(require('./apps/auction'))
 app.use(require('./apps/auction_support'))
 app.use(require('./apps/auction_lots'))
 app.use(require('./apps/artwork_purchase'))
